refactor: 비즈니스 로직을 가지는 객체를 domain 패키지로 변경
