+++ src/main/java/racingcar/domain/CarMoveRule.java
@@ -1,4 +1,4 @@
-package racingcar.model;
+package racingcar.domain;
 
 public interface CarMoveRule {
     boolean isGo();
