+++ src/main/java/racingcar/domain/DefaultCarMoveRule.java
@@ -1,4 +1,4 @@
-package racingcar.model;
+package racingcar.domain;
 
 import racingcar.utils.Random;
