+++ src/main/java/racingcar/domain/RacingGame.java
@@ -1,4 +1,4 @@
-package racingcar.model;
+package racingcar.domain;
 
 public class RacingGame {
