+++ src/test/java/racingcar/domain/RacingGameTest.java
@@ -1,4 +1,4 @@
-package racingcar.model;
+package racingcar.domain;
 
 import java.util.List;
